shooter/requirements.txt
munch
pyyaml